+++ src/js/core/navbar.js
@@ -1,6 +1,6 @@
 import Class from '../mixin/class';
 import FlexBug from '../mixin/flex-bug';
-import {$, $$, addClass, after, assign, css, height, includes, isRtl, isVisible, matches, noop, parents, Promise, query, remove, toFloat, toggleClass, Transition, within} from 'uikit-util';
+import {$, $$, addClass, after, assign, css, height, includes, isRtl, isVisible, matches, noop, Promise, query, remove, toFloat, Transition, within} from 'uikit-util';
 
 export default {
 
@@ -163,7 +163,9 @@ export default {
 
             handler(_, {$el, dir}) {
 
-                toggleClass(this.dropbar, 'uk-navbar-dropbar-slide', this.dropbarMode === 'slide' || parents(this.$el).some(el => css(el, 'position') !== 'static'));
+                if (this.dropbarMode === 'slide') {
+                    addClass(this.dropbar, 'uk-navbar-dropbar-slide');
+                }
 
                 this.clsDrop && addClass($el, `${this.clsDrop}-dropbar`);